ルーティング設定と関数群を分離
